# docs/03_persistence_and_api/openapi.yaml
openapi: 3.1.0
info:
  title: RestoStock API
  description: REST API specification for RestoStock real-time inventory and FEFO traceability system.
...
servers:
  - url: http://localhost:3000
    description: Local Development Server
  - url: https://api.restostock.internal
    description: Production Server

# TK-166 / US-047 / EXT-002 R-01: las 44 operaciones usaban etiquetas que el documento
# nunca declaraba, de modo que todas arrastraban el mismo aviso del linter. Declararlas
# aquí las cierra de una vez y, de paso, da a quien lee el contrato un índice de sus
# áreas funcionales en lugar de siete nombres sueltos repartidos por el archivo.
...
paths:
...
  /api/v1/auth/users:
    get:
      summary: List all operator/admin accounts (ADMIN only, TK-056)
      operationId: listUsers
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users (never includes pinHash)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreateUserResponse'
        '401':
          description: Missing or invalid Bearer JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Authenticated user does not have the ADMIN role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new kitchen operator or admin account (ADMIN only)
      operationId: createUser
      tags:
        - Authentication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Invalid name/operatorCode/pin, or a role that does not exist in the Role catalog (UnknownRoleException, TK-174; the detail lists the valid roles)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid Bearer JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Authenticated user does not have the ADMIN role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >-
            US-051/TK-173: the operatorCode is already taken. Emitted both by the
            use case's fast-path read and by the real unique-index violation, so two
            concurrent creations with the same code get identical treatment (Guard 39).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/users/{id}/status:
    patch:
      summary: Block or reactivate an operator account (ADMIN only)
      operationId: setUserStatus
      tags:
        - Authentication
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserStatusRequest'
      responses:
        '200':
          description: User status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetUserStatusResponse'
        '403':
          description: Authenticated user does not have the ADMIN role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/change-pin:
    post:
      summary: Rotate the PIN of the authenticated operator (US-046 / TK-165)
      description: >-
        Documented in TK-165: the endpoint already existed in code and was missing from this
        contract. Clearing the mandatory first-login rotation (US-046) goes through here, so
        the capability is only complete once it is published.
      operationId: changePin
      tags:
        - Authentication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePinRequest'
      responses:
        '200':
          description: PIN updated; the mandatory rotation flag is cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePinResponse'
        '400':
          description: Invalid PIN format (must be 4-6 digits) or missing userId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: The current PIN is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The account is blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - name
        - operatorCode
        - role
        - pin
      properties:
        name:
          type: string
          example: "Nuevo Operario"
        operatorCode:
          type: string
          pattern: '^[A-Z0-9-]{2,12}$'
          description: >-
            US-051/TK-173: credencial de acceso que el administrador asigna. Se normaliza
            (`trim` + mayúsculas) antes de persistir. Unicidad impuesta por el índice único
            `User_operatorCode_key` de PostgreSQL (Guard 39); un duplicado responde 409.
          example: "CG-01"
        role:
          type: string
          enum: [KITCHEN_STAFF, ADMIN]
        pin:
          type: string
          pattern: '^\d{4,6}$'
          example: "4321"
    CreateUserResponse:
      type: object
      properties:
        id:
          type: string
          description: Clave interna. NO es una credencial — el operario entra con `operatorCode`.
        operatorCode:
          type: string
          example: "CG-01"
        name:
          type: string
        role:
          type: string
          enum: [KITCHEN_STAFF, ADMIN]
        status:
          type: string
          enum: [ACTIVE, BLOCKED]
    SetUserStatusRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [BLOCK, ACTIVATE]
    SetUserStatusResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [ACTIVE, BLOCKED]
...
    ChangePinRequest:
      type: object
      required:
        - userId
        - currentPin
        - newPin
      properties:
        userId:
          type: string
          minLength: 1
        currentPin:
          type: string
          pattern: '^[0-9]{4,6}$'
          description: PIN in force. An incorrect value counts as a failed attempt.
        newPin:
          type: string
          pattern: '^[0-9]{4,6}$'
    ChangePinResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
...
    ErrorResponse:
      description: RFC 7807 Problem Details format (Guard 19)
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
          example: "https://restostock.com/errors/validation-error"
        title:
          type: string
          example: "ValidationError"
        status:
          type: integer
          example: 400
        detail:
          type: string
        instance:
          type: string
          example: "/api/v1/stock/extraction"
...